rename auth client to avoid tss import-protection
