+++ packages/app/src/routes/_authenticated.tsx
@@ -16,8 +16,8 @@ import {
 import { getRequestHeaders } from "@tanstack/react-start/server";
 import { Loader2 } from "lucide-react";
 import { useEffect, useState } from "react";
-import { authClient } from "@/lib/auth.client";
 import { auth } from "@/lib/auth.server";
+import { authClient } from "@/lib/auth-client";
 import { createPartiallyAuthenticatedServerFn } from "@/lib/serverFn";
 
 /**
